demo: one caption implementation, reached through both existing APIs
#253 added `demo-video/caption.ts`, a caption overlay injected into the recorded
page. This branch independently added `AdminUIHelper.caption()`, drawing its own
overlay through Playwright 1.61's `screencast.showOverlay`. Rebasing put both into
`capture.spec.ts` with no textual conflict, so every beat narrated itself twice —
two pills, two near-identical strings, on screen at once. Nothing failed; the video
just looked wrong.

Collapse them into one implementation in `helpers/caption.ts`, which prefers the
screencast overlay and falls back to DOM injection when screencast is unavailable
(Playwright < 1.61, or video off). Screencast is preferred because it composites
into the VIDEO rather than injecting a node into the app being demoed, it survives
full page navigations without re-injection (a DOM node is destroyed on navigate),
and it is the same API `enableDemoCursor()` needs to draw a pointer that tracks
over the admin iframe.

Both entry points stay:

  - `demo-video/caption.ts` re-exports `showCaption`/`clearCaption` unchanged, so
    external demo suites importing that path keep working. The fallback preserves
    `#__hydra_demo_caption__` byte-for-byte, so anything inspecting that node still
    finds it whenever the DOM path is taken.
  - `AdminUIHelper.caption()` / `.clearCaption()` delegate to the same functions
    with the page bound, so mixing both call styles in one demo can no longer
    produce two pills.

`beat()` is now purely the pacing pause; narration opens each beat via
`helper.caption()`. #253's beat phrasing is preserved rather than dropped —
"Switch to block mode" and "Drag blocks to reorder" become two captioned beats
instead of one caption spanning both.
